pyproject.toml
[project]
name = "television"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.10.10",
    "aiohttp-cors>=0.7.0",
    "numpy>=2.1.2",
    "torch>=2.4.1",
    "vuer==0.0.32rc7",
    "pyyaml>=6.0.2",
]
